Bugfix and cleanup.

+++ ocl.c
@@ -782,10 +782,10 @@ built:
 	 * 2 greater >= required amount earlier */
 	if (bufsize > cgpu->max_alloc) {
 		applog(LOG_WARNING, "Maximum buffer memory device %d supports says %lu",
-			   gpu, (long unsigned int)(cgpu->max_alloc));
-		applog(LOG_WARNING, "Your scrypt settings come to %d", (int)bufsize);
+			   gpu, (unsigned long)(cgpu->max_alloc));
+		applog(LOG_WARNING, "Your scrypt settings come to %lu", (unsigned long)bufsize);
 	}
-	applog(LOG_DEBUG, "Creating scrypt buffer sized %d", (int)bufsize);
+	applog(LOG_DEBUG, "Creating scrypt buffer sized %lu", (unsigned long)bufsize);
 	clState->padbufsize = bufsize;
 
 	/* This buffer is weird and might work to some degree even if
